test:  Member 인수테스트 공통코드 분리
